requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Django==6.0.1
djangorestframework==3.16.1
filelock==3.20.3
fsspec==2026.1.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.1
packaging==25.0
psycopg2-binary==2.9.11
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.0
setuptools==80.9.0
sqlparse==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
